+++ tests/resources/runners_test/parameters.py
@@ -10,10 +10,10 @@ control_structure = {
         SimulationInstruction(
             conditions=[TimePoints([1, 2, 3, 4])],
             events=Sequence([
-                Event(toy_inc, parameters={"incrementation": 2})
+                Event(toy_inc, static_parameters={"incrementation": 2})
             ])
         )
     ],
     "transition": ToyTransition(),
-    "end_condition": Condition[ToyModel](lambda x: x.time > 4)
+    "end_condition": Condition[ToyModel](lambda x: x.computational_unit.time > 4)
 }