+++ tests/resources/runners_test/parameters_branching.py
@@ -2,7 +2,7 @@ from lukefi.metsi.domain.conditions import TimePoints
 from lukefi.metsi.sim.condition import Condition
 from lukefi.metsi.sim.simulation_instruction import SimulationInstruction
 from lukefi.metsi.sim.generators import Alternatives, Sequence, Event
-from lukefi.metsi.sim.operations import do_nothing
+from lukefi.metsi.sim.treatment import do_nothing
 from tests.toy_model import ToyModel, ToyTransition, toy_inc
 
 
@@ -14,12 +14,12 @@ control_structure = {
                 Sequence([Event(do_nothing)]),
                 Alternatives([
                     Event(do_nothing),
-                    Event(toy_inc, parameters={"incrementation": 1}),
-                    Event(toy_inc, parameters={"incrementation": 2})
+                    Event(toy_inc, static_parameters={"incrementation": 1}),
+                    Event(toy_inc, static_parameters={"incrementation": 2})
                 ])
             ])
         )
     ],
     "transition": ToyTransition(),
-    "end_condition": Condition[ToyModel](lambda x: x.time > 2)
+    "end_condition": Condition[ToyModel](lambda x: x.computational_unit.time > 2)
 }